% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildGLMMF}
\alias{buildGLMMF}
\title{Build Generalized Linear Mixed-Effects Model with Latent Factors}
\usage{
buildGLMMF(group, response, common.fixed, distinct.fixed, random,
  nfactors = 0, data, distribution = c("gaussian", "poisson", "binomial",
  "gamma", "negative binomial"), u, random.cov, factors,
  tol = .Machine$double.eps^0.5)
}
\arguments{
\item{group}{Name of the grouping variable in data.
Only one grouping variable is allowed and the group sizes must be equal. In case of unequal group sizes,
patch missing rows with NA's.}

\item{response}{Name of the response variable in data.}

\item{common.fixed}{formula for common fixed effects. LHS of the formula is ignored if present.}

\item{distinct.fixed}{Formula for distinct fixed effects i.e. each group has separate regression
coefficient. This formula cannot contain variables which are already present in
\code{common.fixed} or \code{random} formulas, as in that case the model would not be identifiable.
LHS of the formula is ignored if present.}

\item{random}{Formula for random effects. LHS of the formula is ignored if present.}

\item{data}{Data frame containing the variables in the model. Must contain all variables used formulas and
variables defined in \code{group} and \code{response}}

\item{distribution}{Character. Possible values are "gaussian", "poisson",
"binomial", "negative binomial" and "gamma". Default is "gaussian".}
}
\description{
Function \code{buildGLMMF} builds GLMM with latent factors using methods based on state space modelling.
}

