% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallel_glmmf}
\alias{parallel_glmmf}
\title{Parallel estimation of GLMMF}
\usage{
parallel_glmmf(group, response, common.fixed, distinct.fixed, random,
  nfactors = 0, data, distribution = c("gaussian", "poisson", "binomial",
  "gamma", "negative binomial"), u, correlating.effects = TRUE,
  common.dispersion = TRUE, init.random.cov, init.dispersion, init.factor,
  init.theta, return.model = TRUE, maxiter = 50, maxiter2 = 10,
  convtol = 1e-08, estimate = TRUE, tol = .Machine$double.eps^0.5,
  trace = 0, seed = 1, group_size = 5, threads = 1, samples = 1,
  verbose = FALSE, ...)
}
\arguments{
\item{group}{Name of the grouping variable in data.
Only one grouping variable is allowed and the group sizes must be equal. In case of unequal group sizes,
patch missing rows with NA's.}

\item{response}{Name of the response variable in data.}

\item{common.fixed}{formula for common fixed effects. LHS of the formula is ignored if present.}

\item{distinct.fixed}{Formula for distinct fixed effects i.e. each group has separate regression
coefficient. This formula cannot contain variables which are already present in
\code{common.fixed} or \code{random} formulas, as in that case the model would not be identifiable.
LHS of the formula is ignored if present.}

\item{random}{Formula for random effects. LHS of the formula is ignored if present.}

\item{data}{Data frame containing the variables in the model. Must contain all variables used formulas and
variables defined in \code{group} and \code{response}}

\item{distribution}{Distribution of observations. Possible choices are "gaussian", "poisson", "binomial", "gamma and
"negative binomial". Default is "gaussian".}

\item{correlating.effects}{Logical. Default is TRUE.}

\item{init.dispersion}{Initial values for dispersion paremeters for Gaussian, negative binomial and Gamma distributions.}

\item{maxiter}{Integer. Number of iterations for in iterative weighted least squares.}

\item{init.random}{Initial values for random effect covariances.}
}
\description{
Function \code{parallel_glmmf} is a parallel version of glmmf function.
}

